Arreglo de estructura por (duplicación de carpetas)
