+++ src/TaskFlow.Console/Services/TaskService.cs
@@ -49,14 +49,46 @@ namespace TaskFlow.Services
 
         // Cambiar estado
         public bool UpdateTaskStatus(int id, TaskStatus newStatus) 
-        { 
-            return false; 
-        }
+        {
+            // 1. Buscamos la tarea por su ID usando un foreach tradicional
+            TaskItem taskFound = null;
 
+            foreach (TaskItem task in _tasks)
+            {
+                if (task.Id == id)
+                {
+                    taskFound = task;
+                    break; // Cortamos el bucle porque ya la encontramos
+                }
+            }
+
+            // 2. Verificamos si encontramos la tarea
+            if (taskFound != null)
+            {
+                // Actualizamos el estado
+                taskFound.Status = newStatus;
+                // Registramos la fecha exacta de la modificación
+                taskFound.UpdatedAt = DateTime.Now;
+
+                return true; // Indicamos que la operación fue un éxito
+            }
+            else
+            {
+                return false; // Indicamos que no se encontró el ID
+            }
+        }
+    
         // Validar si hay tareas
         public bool HasTasks() 
-        { 
-            return false; 
+        {
+            if (_tasks.Count > 0)
+            {
+                return true;
+            }
+            else
+            {
+                return false;
+            }
         }
 
         // Cambiar responsable
